+++ apps/bvp/stagnation.cpp
@@ -116,7 +116,7 @@ AxiStagnBVP::AxiStagnBVP(igthermo_t* ph, int nsp, int points) :
         m_refiner->setActive(3, false);
 
         vector_fp gr;
-        for (int ng = 0; ng < m_points; ng++) gr.push_back(1.0*ng/m_points);
+        for (size_t ng = 0; ng < m_points; ng++) gr.push_back(1.0*ng/m_points);
         setupGrid(m_points, DATA_PTR(gr));
         setID("stagnation flow");
     }
@@ -221,7 +221,7 @@ AxiStagnBVP::AxiStagnBVP(igthermo_t* ph, int nsp, int points) :
         m_thermo->setTemperature(0.5*(T(x,j)+T(x,j+1)));
         const doublereal* yyj = x + m_nv*j + c_offset_Y;
         const doublereal* yyjp = x + m_nv*(j+1) + c_offset_Y;
-        for (int k = 0; k < m_nsp; k++)
+        for (size_t k = 0; k < m_nsp; k++)
             m_ybar[k] = 0.5*(yyj[k] + yyjp[k]);
         m_thermo->setMassFractions_NoNorm(DATA_PTR(m_ybar));
         m_thermo->setPressure(m_press);
